Move getThemeShowcaseDescription
