+++ client/my-sites/themes/theme-showcase.jsx
@@ -24,7 +24,6 @@ import { getSiteSlug } from 'state/sites/selectors';
 import { getCurrentUserId } from 'state/current-user/selectors';
 import ThemePreview from './theme-preview';
 import config from 'config';
-import getThemeShowcaseDescription from 'state/selectors/get-theme-showcase-description';
 import getThemeShowcaseTitle from 'state/selectors/get-theme-showcase-title';
 import prependThemeFilterKeys from 'state/selectors/prepend-theme-filter-keys';
 import { recordTracksEvent } from 'state/analytics/actions';
@@ -39,6 +38,7 @@ import {
 	getActiveTheme,
 	getThemeFilterTerms,
 	getThemeFilterToTermTable,
+	getThemeShowcaseDescription,
 } from 'state/themes/selectors';
 import UpworkBanner from 'blocks/upwork-banner';
 import RecommendedThemes from './recommended-themes';
