+++ client/state/themes/selectors/get-theme-showcase-description.js
@@ -6,11 +6,12 @@ import { get, includes } from 'lodash';
 /**
  * Internal dependencies
  */
-import { findThemeFilterTerm, getThemeFilterTerm } from 'state/themes/selectors';
+import { findThemeFilterTerm } from 'state/themes/selectors/find-theme-filter-term';
+import { getThemeFilterTerm } from 'state/themes/selectors/get-theme-filter-term';
 
 import 'state/themes/init';
 
-export default function getThemeShowcaseDescription( state, { filter, tier, vertical } = {} ) {
+export function getThemeShowcaseDescription( state, { filter, tier, vertical } = {} ) {
 	if ( vertical ) {
 		const description = get( getThemeFilterTerm( state, 'subject', vertical ), 'description' );
 		if ( description ) {
